+++ packages/main/package.json
@@ -48,7 +48,7 @@
     "build": "npm-run-all --sequential clean lint build:templates build:i18n-exports build:samples build:styles copy:src build:bundle copy:i18n copy:webcomponents-polyfill ",
     "build:bundle": "rollup -c --environment ES5_BUILD",
     "build:templates": "mkdirp dist/generated/templates && node ./lib/hbs2ui5/index.js -d src/ -o dist/generated/templates",
-    "build:i18n-exports": "mkdirp dist/generated/i18n && node ./lib/i18n-transform/index.js src/i18n/messagebundle.properties dist/generated/i18n",
+    "build:i18n-exports": "mkdirp dist/generated/i18n && node ./lib/i18n-transform/index.js src/i18n dist/generated/i18n",
     "build:samples": "npm run copy:test --scripts-prepend-node-path && npm run build:api --scripts-prepend-node-path && npm run build:docs --scripts-prepend-node-path && node lib/playground/index.js",
     "build:pages": "npm run copy:pages",
     "build:qunit": "npm run copy:qunit",
