+++ packages/main/src/i18n/messagebundle_en.properties
@@ -1,0 +1,65 @@
+#This is the resource bundle for the UI5 Web Components
+#__ldi.translation.uuid=96bea51a-d5e3-46f0-b1d1-514d97be02ec
+
+#XBUT: Card Content aria-label text
+ARIA_LABEL_CARD_CONTENT=Card Content
+
+#XBUT: Card aria-roledescription text
+ARIA_ROLEDESCRIPTION_CARD=Card
+
+#XACT: ARIA announcement for the Avatar default tooltip
+AVATAR_TOOLTIP=Avatar
+
+#XACT: ARIA announcement for the badge
+BADGE_DESCRIPTION=Badge
+
+#XTOL: text that could be show if BusyIndicator is active
+BUSY_INDICATOR_TITLE=Please wait
+
+#XACT: ARIA announcement for the accept button
+BUTTON_ARIA_TYPE_ACCEPT=Positive Action
+
+#XACT: ARIA announcement for the reject button
+BUTTON_ARIA_TYPE_REJECT=Negative Action
+
+#XACT: ARIA announcement for the emphasized button
+BUTTON_ARIA_TYPE_EMPHASIZED=Emphasized
+
+#XBUT: A link that can be clicked to display more/all items
+INPUT_SUGGESTIONS=Suggestions available
+
+#XFLD: Subtle link description label
+LINK_SUBTLE=Subtle
+
+#XFLD: Emphasized link description label
+LINK_EMPHASIZED=Emphasized
+
+#XTOL: Tooltip of messgae strip close button
+MESSAGE_STRIP_CLOSE_BUTTON=Message Strip Close
+
+#XFLD: Token number indicator which is used to show more tokens in Tokenizers inside MultiInput and MultiComboBox
+MULTIINPUT_SHOW_MORE_TOKENS={0} More
+
+#XTOL: Tooltip for panel expand title
+PANEL_ICON=Expand/Collapse
+
+#XACT: ARIA announcement for the switch on
+SWITCH_ON=On
+
+#XACT: ARIA announcement for the switch off
+SWITCH_OFF=Off
+
+#XTXT: Text for characters left
+TEXTAREA_CHARACTERS_LEFT={0} characters remaining
+
+#XTXT: Text for characters over
+TEXTAREA_CHARACTERS_EXCEEDED={0} characters over limit
+
+#XTOL: text that is appended to the tooltips of input fields etc. which are marked to have an error
+VALUE_STATE_ERROR=Invalid entry
+
+#XTOL: text that is appended to the tooltips of input fields etc. which are marked to have a warning
+VALUE_STATE_WARNING=Warning issued
+
+#XTOL: text that is appended to the tooltips of input fields etc. which are marked to be in success state
+VALUE_STATE_SUCCESS=Entry successfully validated
